make 6g constants behave as ken proposes.  (i hope.) various bug fixes and tests involving constants.
test/const1.go is the major new test case.

R=ken
OCL=26216
CL=26224

+++ src/lib/syscall/errstr_darwin.go
@@ -235,7 +235,7 @@ func str(val int64) string {  // do it here rather than with fmt to avoid depend
 }
 
 func Errstr(errno int64) string {
-	if errno < 0 || errno >= len(error) {
+	if errno < 0 || errno >= int64(len(error)) {
 		return "Error " + str(errno)
 	}
 	return error[errno]
